refactor(history): split page orchestration helpers
