+++ lib/history/history_page.dart
@@ -1,30 +1,21 @@
-import 'dart:async';
 import 'package:flutter/material.dart';
 import 'package:flutter_hooks/flutter_hooks.dart';
 import 'package:hooks_riverpod/hooks_riverpod.dart';
-import 'package:threed_print_cost_calculator/core/analytics/app_analytics.dart';
-import 'package:threed_print_cost_calculator/history/components/history_export_options_sheet.dart';
-import 'package:threed_print_cost_calculator/history/components/history_export_preview_sheet.dart';
 import 'package:threed_print_cost_calculator/history/components/history_overflow_hint.dart';
 import 'package:threed_print_cost_calculator/l10n/app_localizations.dart';
-import 'package:threed_print_cost_calculator/purchases/paywall_presenter.dart';
 import 'package:threed_print_cost_calculator/purchases/premium_access_providers.dart';
-import 'package:threed_print_cost_calculator/purchases/premium_upsell_helper.dart';
 import 'package:threed_print_cost_calculator/shared/providers/app_providers.dart';
-import 'package:threed_print_cost_calculator/shared/utils/csv_utils.dart';
-import 'provider/history_paged_notifier.dart';
 
 import 'components/history_empty_state.dart';
-import 'components/history_teaser.dart';
 import 'components/history_list_view.dart';
 import 'components/history_search_bar.dart';
+import 'components/history_teaser.dart';
+import 'hooks/history_overflow_hint.dart';
+import 'hooks/history_page_actions.dart';
 import 'hooks/history_search_query.dart';
+import 'provider/history_paged_notifier.dart';
 
 enum HistoryPageMode { full, teaser }
-
-const _overflowHintPreferenceKey = 'history_overflow_hint_seen_v2';
-const _overflowMenuOpenedPreferenceKey = 'history_overflow_menu_opened_v1';
-const _overflowHintDuration = Duration(seconds: 4);
 
 class HistoryPage extends HookConsumerWidget {
   const HistoryPage({
@@ -43,17 +34,17 @@ class HistoryPage extends HookConsumerWidget {
     final paged = ref.watch(historyPagedProvider);
     final policy = ref.watch(premiumAccessPolicyProvider);
 
+    final actions = HistoryPageActions(ref: ref);
+
     if (mode == HistoryPageMode.teaser) {
       return HistoryTeaser(
-        onUpgradePressed: () => _showTeaserPaywall(
+        onUpgradePressed: () => actions.showTeaserPaywall(
           context,
-          ref: ref,
           isPremium: policy.isPremium,
           source: 'history_teaser_primary',
         ),
-        onExportPreviewPressed: () => _showTeaserPreview(
+        onExportPreviewPressed: () => actions.showTeaserPreview(
           context,
-          ref: ref,
           isPremium: policy.isPremium,
           source: 'history_teaser_secondary',
         ),
@@ -68,67 +59,15 @@ class HistoryPage extends HookConsumerWidget {
         ? paged.items.take(historyLimit).toList(growable: false)
         : paged.items;
 
-    final prefs = ref.read(sharedPreferencesProvider);
     final controller = useTextEditingController(text: paged.query);
     final scrollController = useScrollController();
-    final showOverflowHint = useState(false);
-    final overflowHintTimer = useRef<Timer?>(null);
+    final overflowHint = useHistoryOverflowHint(
+      ref: ref,
+      itemCount: paged.items.length,
+    );
 
     useHistorySearchQuery(ref: ref, controller: controller);
 
-    Future<void> markOverflowHintSeen() async {
-      if (prefs.getBool(_overflowHintPreferenceKey) == true) return;
-      await prefs.setBool(_overflowHintPreferenceKey, true);
-    }
-
-    Future<void> markOverflowMenuOpened() async {
-      if (prefs.getBool(_overflowMenuOpenedPreferenceKey) == true) return;
-      await prefs.setBool(_overflowMenuOpenedPreferenceKey, true);
-      AppAnalytics.safeLog(() => AppAnalytics.log('history_overflow_opened'));
-    }
-
-    void dismissOverflowHint() {
-      if (!showOverflowHint.value) return;
-      overflowHintTimer.value?.cancel();
-      showOverflowHint.value = false;
-    }
-
-    useEffect(() {
-      var disposed = false;
-
-      Future<void> maybeShowOverflowHint() async {
-        if (paged.items.isEmpty) return;
-
-        final hasSeenHint = prefs.getBool(_overflowHintPreferenceKey) ?? false;
-        final hasOpenedMenu =
-            prefs.getBool(_overflowMenuOpenedPreferenceKey) ?? false;
-        if (hasSeenHint || hasOpenedMenu || showOverflowHint.value) return;
-
-        await markOverflowHintSeen();
-        AppAnalytics.safeLog(
-          () => AppAnalytics.log('history_overflow_hint_shown'),
-        );
-        if (disposed) return;
-
-        showOverflowHint.value = true;
-        overflowHintTimer.value?.cancel();
-        overflowHintTimer.value = Timer(_overflowHintDuration, () {
-          if (!disposed) {
-            showOverflowHint.value = false;
-          }
-        });
-      }
-
-      unawaited(maybeShowOverflowHint());
-
-      return () {
-        disposed = true;
-        overflowHintTimer.value?.cancel();
-      };
-    }, [paged.items.length]);
-
-    // Load first page on mount — schedule after the first frame to avoid
-    // modifying providers during the widget build lifecycle.
     useEffect(() {
       WidgetsBinding.instance.addPostFrameCallback((_) {
         ref.read(historyPagedProvider.notifier).refreshIfNeeded();
@@ -136,15 +75,11 @@ class HistoryPage extends HookConsumerWidget {
       return null;
     }, [appRefreshTick]);
 
-    // Infinite scroll: listen to scroll controller and load more when near bottom
     useEffect(() {
       void scrollListener() {
-        final sc = scrollController;
-        if (!sc.hasClients) return;
-        final max = sc.position.maxScrollExtent;
-        final current = sc.position.pixels;
-
-        // Trigger when within 200px of bottom
+        if (!scrollController.hasClients) return;
+        final max = scrollController.position.maxScrollExtent;
+        final current = scrollController.position.pixels;
         const threshold = 200.0;
         if (current >= (max - threshold)) {
           final notifier = ref.read(historyPagedProvider.notifier);
@@ -185,14 +120,14 @@ class HistoryPage extends HookConsumerWidget {
                 child: HistorySearchBar(
                   controller: controller,
                   onExportPressed: policy.bulkHistoryExport().allowed
-                      ? () => _showExportOptions(context, ref)
+                      ? () => actions.showExportOptions(context)
                       : null,
                 ),
               ),
               SliverToBoxAdapter(
                 child: AnimatedSwitcher(
                   duration: const Duration(milliseconds: 180),
-                  child: showOverflowHint.value
+                  child: overflowHint.showOverflowHint.value
                       ? HistoryOverflowHint(
                           key: const ValueKey<String>('history.overflow.hint'),
                           message: l10n.historyOverflowHint,
@@ -210,8 +145,8 @@ class HistoryPage extends HookConsumerWidget {
                   items: visibleItems,
                   onHistoryLoaded: onHistoryLoaded,
                   onOverflowMenuOpened: () async {
-                    await markOverflowMenuOpened();
-                    dismissOverflowHint();
+                    await overflowHint.markOverflowMenuOpened();
+                    overflowHint.dismissOverflowHint();
                   },
                 ),
               SliverToBoxAdapter(
@@ -243,114 +178,5 @@ class HistoryPage extends HookConsumerWidget {
         },
       ),
     );
-  }
-
-  void _showExportOptions(BuildContext context, WidgetRef ref) {
-    final l10n = AppLocalizations.of(context)!;
-
-    showModalBottomSheet<void>(
-      context: context,
-      builder: (_) {
-        return HistoryExportOptionsSheet(
-          onExportSelected: (range) async {
-            await _exportHistoryRange(ref, l10n, range);
-          },
-        );
-      },
-    );
-  }
-
-  Future<void> _exportHistoryRange(
-    WidgetRef ref,
-    AppLocalizations l10n,
-    ExportRange range,
-  ) async {
-    final policy = ref.read(premiumAccessPolicyProvider);
-    if (!await requirePremium(
-      ref.read(paywallPresenterProvider),
-      policy.bulkHistoryExport(),
-      purchaseSource: 'history_export',
-      recheck: () => Future.value(
-        ref.read(premiumAccessPolicyProvider).bulkHistoryExport().allowed,
-      ),
-    )) {
-      return;
-    }
-
-    // Use mixed history export to handle both single-print and batch quotes
-    await ref
-        .read(csvUtilsProvider)
-        .exportMixedHistoryForRange(
-          range,
-          shareText: l10n.mixedHistoryExportShareText,
-        );
-    AppAnalytics.safeLog(() => AppAnalytics.exportUsed('history'));
-  }
-
-  Future<void> _showTeaserPreview(
-    BuildContext context, {
-    required WidgetRef ref,
-    required bool isPremium,
-    required String source,
-  }) async {
-    AppAnalytics.safeLog(
-      () => AppAnalytics.premiumFeatureTapped(
-        'history',
-        isPro: isPremium,
-        source: source,
-      ),
-    );
-
-    final l10n = AppLocalizations.of(context)!;
-    final csvPreview = [
-      l10n.historyCsvHeader,
-      '"Benchy",19.25,12.50,3.00,2.50,123,06:20',
-      '"Prusa MK4S",24.10,15.75,3.40,2.75,142,05:10',
-    ].join('\n');
-
-    if (!context.mounted) return;
-
-    await showModalBottomSheet<void>(
-      context: context,
-      isScrollControlled: true,
-      showDragHandle: true,
-      builder: (sheetContext) {
-        return HistoryExportPreviewSheet(
-          csvPreview: csvPreview,
-          onDownloadPressed: () async {
-            Navigator.of(sheetContext).pop();
-            await _showTeaserPaywall(
-              context,
-              ref: ref,
-              isPremium: isPremium,
-              source: source,
-            );
-          },
-        );
-      },
-    );
-  }
-
-  Future<void> _showTeaserPaywall(
-    BuildContext context, {
-    required WidgetRef ref,
-    required bool isPremium,
-    required String source,
-  }) async {
-    AppAnalytics.safeLog(
-      () => AppAnalytics.premiumFeatureTapped(
-        'history',
-        isPro: isPremium,
-        source: source,
-      ),
-    );
-    await ref
-        .read(paywallPresenterProvider)
-        .present(
-          'pro',
-          triggerFeature: 'history',
-          purchaseSource: source,
-          source: source,
-        );
   }
 }
